requirements.txt
instabot==0.117.0
python-dotenv==1.0.1
Requests==2.32.3
setuptools==68.2.0
facebook-sdk
pandas~=2.2.2